fix: some minor fixes

+++ src/core/tests/test-va.c
@@ -6,7 +6,7 @@ int main(void)
     IC_ASSERT(strcmp(t1, "this is a test") == 0);
 
     char max[16384];
-    for (int i = 0; i < sizeof(max); i++)
+    for (size_t i = 0; i < sizeof(max); i++)
         max[i] = 'a';
 
     char *test = va("%s", max);
